requirements.txt
django==2.1.3
djangorestframework==3.9.0
django-filter
django-cors-headers==2.4.0
gunicorn
django-heroku
psycopg2-binary
